[WO-37] z uzyciem psutil psutil.pid_exists

+++ utrzymanie_wersji.py
@@ -11,6 +11,7 @@ import zipfile
 from dotenv import load_dotenv
 import json
 import signal
+import psutil
 
 def nazwa_programu():
     return "update_projektu_skryptu_klraspi.py"
@@ -171,9 +172,11 @@ def sprawdz_czy_skrypty_klraspi_dziala_i_ubij_jesli_dziala(basic_path_ram):
             linie=file.readline()
         numer_pid=int(linie)
         if os.name == "posix":
+
             if int(numer_pid)>-1:
-                os.kill(int(numer_pid), signal.SIGTERM)
-                os.remove(skrypty_klraspi_path)
+                if psutil.pid_exists(numer_pid) == True:
+                    os.kill(int(numer_pid), signal.SIGTERM)
+                    os.remove(skrypty_klraspi_path)
         else:
             raise ExceptionWindows
